refactor(LandingView.vue): LandingView split into Playlist、Openbutton and LandingView
LandingView split into two components : Playlist and Openbutton ,use $bus to transfer the value
