+++ package.json
@@ -127,6 +127,7 @@
     "copy-webpack-plugin": "^4.0.1",
     "cross-env": "^5.1.6",
     "css-loader": "^0.28.11",
+    "cz-conventional-changelog": "^2.1.0",
     "del": "^3.0.0",
     "devtron": "^1.4.0",
     "electron-builder": "^20.19.2",
@@ -171,5 +172,10 @@
     "webpack-dev-server": "^2.7.1",
     "webpack-hot-middleware": "^2.22.2",
     "webpack-merge": "^4.1.0"
+  },
+  "config": {
+    "commitizen": {
+      "path": "./node_modules/cz-conventional-changelog"
+    }
   }
 }
